+++ Sources/DataLoader/OAuth2DataLoader.swift
@@ -37,9 +37,20 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	public var alsoIntercept403: Bool = false
 	
 	
-	public init(oauth2: OAuth2) {
+	/**
+	Designated initializer.
+	
+	Provide `host` if 301 and 302 redirects should be followed automatically, as long as they appear on the same host.
+	
+	- parameter oauth2: The OAuth2 instance to use for authorization when loading data.
+	- parameter host:   If given will handle redirects within the same host by way of `OAuth2DataLoaderSessionTaskDelegate`
+	*/
+	public init(oauth2: OAuth2, host: String? = nil) {
 		self.oauth2 = oauth2
 		super.init(logger: oauth2.logger)
+		if let host = host {
+			sessionDelegate = OAuth2DataLoaderSessionTaskDelegate(loader: self, host: host)
+		}
 	}
 	
 	
@@ -114,6 +125,7 @@ open class OAuth2DataLoader: OAuth2Requestable {
 			catch OAuth2Error.unauthorizedClient {
 				if retry {
 					self.enqueue(request: request, callback: callback)
+					self.oauth2.clientConfig.accessToken = nil
 					self.attemptToAuthorize() { json, error in
 						
 						// dequeue all if we're authorized, throw all away if something went wrong
@@ -129,7 +141,7 @@ open class OAuth2DataLoader: OAuth2Requestable {
 					callback(response)
 				}
 			}
-				
+			
 			// some other error, pass along
 			catch {
 				callback(response)
@@ -138,12 +150,12 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	}
 	
 	/**
-	If not already authorizing, will use its `oauth2` instance to start authorization.
+	If not already authorizing, will use its `oauth2` instance to start authorization after forgetting any tokens (!).
 	
 	This method will ignore calls while authorization is ongoing, meaning you will only get the callback once per authorization cycle.
 	
 	- parameter callback: The callback passed on from `authorize(callback:)`. Authorization finishes successfully (auth parameters will be
-	                      non-nil but may be an empty dict), fails (error will be non-nil) or is cancelled (both params and error are nil)
+	                      non-nil but may be an empty dict), fails (error will be non-nil) or is canceled (both params and error are nil)
 	*/
 	open func attemptToAuthorize(callback: @escaping ((OAuth2JSON?, OAuth2Error?) -> Void)) {
 		if !isAuthorizing {
@@ -201,8 +213,13 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	func retryAll() {
 		dequeueAndApply() { req in
 			var request = req.request
-			request.sign(with: oauth2)
-			self.perform(request: request, retry: false, callback: req.callback)
+			do {
+				try request.sign(with: oauth2)
+				self.perform(request: request, retry: false, callback: req.callback)
+			}
+			catch let error {
+				NSLog("OAuth2.DataLoader.retryAll(): \(error)")
+			}
 		}
 	}